requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.4.0
chromadb>=0.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
tiktoken>=0.5.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
unstructured>=0.10.0